chore: migrate Callout to TypeScript
