+++ docs/Callout.md
@@ -1,15 +1,15 @@
-<!-- This file was autogenerated from Callout.js do not modify -->
+<!-- This file was autogenerated from Callout.tsx do not modify -->
 # <MapboxGL.Callout />
 Callout that displays information about a selected annotation near the annotation.
 
 ## props
 | Prop | Type | Default | Required | Description |
 | ---- | :-- | :----- | :------ | :---------- |
-| title | `string` | `none` | `false` | String that gets displayed in the default callout. |
-| style | `any` | `none` | `false` | Style property for the Animated.View wrapper, apply animations to this |
-| containerStyle | `any` | `none` | `false` | Style property for the native RCTMGLCallout container, set at your own risk. |
-| contentStyle | `any` | `none` | `false` | Style property for the content bubble. |
-| tipStyle | `any` | `none` | `false` | Style property for the triangle tip under the content. |
-| textStyle | `any` | `none` | `false` | Style property for the title in the content bubble. |
+| title | `string` | `none` | `true` | String that gets displayed in the default callout. |
+| style | `ViewStyle` | `none` | `false` | Style property for the Animated.View wrapper, apply animations to this |
+| containerStyle | `ViewStyle` | `none` | `false` | Style property for the native RCTMGLCallout container, set at your own risk. |
+| contentStyle | `ViewStyle` | `none` | `false` | Style property for the content bubble. |
+| tipStyle | `ViewStyle` | `none` | `false` | Style property for the triangle tip under the content. |
+| textStyle | `ViewStyle` | `none` | `false` | Style property for the title in the content bubble. |
