+++ javascript/components/Callout.tsx
@@ -1,14 +1,14 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { ReactNode } from 'react';
 import {
   View,
   Text,
   Animated,
   requireNativeComponent,
   StyleSheet,
+  ViewStyle,
+  ViewProps,
+  StyleProp,
 } from 'react-native';
-
-import { viewPropTypes } from '../utils';
 
 export const NATIVE_MODULE_NAME = 'RCTMGLCallout';
 
@@ -48,60 +48,68 @@ const styles = StyleSheet.create({
   },
 });
 
+type NativeProps = {
+  children: ReactNode;
+  style: StyleProp<ViewStyle>;
+};
+
+type Props = Omit<ViewProps, 'style'> & {
+  /**
+   * String that gets displayed in the default callout.
+   */
+  title: string;
+
+  /**
+   * Style property for the Animated.View wrapper, apply animations to this
+   */
+  style?: ViewStyle;
+
+  /**
+   * Style property for the native RCTMGLCallout container, set at your own risk.
+   */
+  containerStyle?: ViewStyle;
+
+  /**
+   * Style property for the content bubble.
+   */
+  contentStyle?: ViewStyle;
+
+  /**
+   * Style property for the triangle tip under the content.
+   */
+  tipStyle?: ViewStyle;
+
+  /**
+   * Style property for the title in the content bubble.
+   */
+  textStyle?: ViewStyle;
+};
+
 /**
  *  Callout that displays information about a selected annotation near the annotation.
  */
-class Callout extends React.PureComponent {
-  static propTypes = {
-    ...viewPropTypes,
-
-    /**
-     * String that gets displayed in the default callout.
-     */
-    title: PropTypes.string,
-
-    /**
-     * Style property for the Animated.View wrapper, apply animations to this
-     */
-    style: PropTypes.any,
-
-    /**
-     * Style property for the native RCTMGLCallout container, set at your own risk.
-     */
-    containerStyle: PropTypes.any,
-
-    /**
-     * Style property for the content bubble.
-     */
-    contentStyle: PropTypes.any,
-
-    /**
-     * Style property for the triangle tip under the content.
-     */
-    tipStyle: PropTypes.any,
-
-    /**
-     * Style property for the title in the content bubble.
-     */
-    textStyle: PropTypes.any,
-  };
-
+class Callout extends React.PureComponent<Props> {
   get _containerStyle() {
-    return [
+    const style = [
       {
         position: 'absolute',
         zIndex: 999,
         backgroundColor: 'transparent',
-      },
-      this.props.containerStyle,
+      } as ViewStyle,
     ];
+
+    if (this.props.containerStyle) {
+      style.push(this.props.containerStyle);
+    }
+
+    return style;
   }
 
   get _hasChildren() {
     return React.Children.count(this.props.children) > 0;
   }
 
-  _renderDefaultCallout() {
+  _renderDefaultCallout(): ReactNode {
     return (
       <Animated.View style={[styles.container, this.props.style]}>
         <View style={[styles.content, this.props.contentStyle]}>
@@ -114,7 +122,7 @@ class Callout extends React.PureComponent {
     );
   }
 
-  _renderCustomCallout() {
+  _renderCustomCallout(): ReactNode {
     return (
       <Animated.View {...this.props} style={this.props.style}>
         {this.props.children}
@@ -134,6 +142,6 @@ class Callout extends React.PureComponent {
   }
 }
 
-const RCTMGLCallout = requireNativeComponent(NATIVE_MODULE_NAME, Callout);
+const RCTMGLCallout = requireNativeComponent<NativeProps>(NATIVE_MODULE_NAME);
 
 export default Callout;
